Move root page
