+++ tests/routers/test_pages.py
@@ -1,0 +1,9 @@
+import pytest
+from httpx import AsyncClient
+
+
+@pytest.mark.anyio
+async def test_root(client: AsyncClient):
+    response = await client.get("/")
+    assert response.status_code == 200
+    assert "Meal Planner" in response.text
